+++ package.json
@@ -13,6 +13,7 @@
     "@tanstack/react-query-devtools": "^5.29.0",
     "axios": "^1.6.8",
     "immer": "^10.0.4",
+    "lodash": "^4.17.21",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
